+++ changelog.md
@@ -8,6 +8,7 @@
   `BOTH.contains(DOWNSTREAM)` or `DOWNSTREAM in BOTH`
 * Removed deprecated function `NetworkConsumerClient.getFeeder`.
 * Renamed function 'breadFirstSupplier' to 'breadthFirstSupplier' in 'BasicQueue' class.
+* Updated the AssignToFeeders to support assigning AuxiliaryEquipment.
 
 ### New Features